[memprof] Move YAML support to MemProfYAML.h (NFC)
The YAML support is increasing in size, so this patch moves it to a
separate file.
